requirements.txt
Django==1.8.6
django-admin-sortable==2.0.7
django-mptt==0.7.4
django-multiselectfield==0.1.3
django-tagging==0.4
django-tinymce==2.0.6
djangorestframework==3.3.1
djangular==0.2.7
feedparser==5.2.1
gunicorn==19.3.0
jsonfield==1.0.3
Markdown==2.6.4
Pillow==3.0.0
psycopg2==2.6.1
pytz==2015.7
simplejson==3.8.1
wheel==0.24.0
